+++ CHANGELOG.md
@@ -13,4 +13,6 @@ All notable changes to this project will be documented in this file.
 - Changed default panic message to be more professional
 - Relicensed from MIT to MIT/Apache-2.0 dual licensing
 - Moved `get_verbosity` -> `Verbosity::from_env`
+- Fix readability on light terminal themes
+- Fix deadlock when unwrapping an error while printing the panic
 - Many internal tweaks